Cambio guardado de imágenes

+++ database/migrations/2024_07_05_065856_add_image_to_rooms_table.php
@@ -4,25 +4,33 @@ use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-return new class extends Migration
+class AddImageToRoomsTable extends Migration
 {
     /**
      * Run the migrations.
+     *
+     * @return void
      */
-    public function up(): void
+    public function up()
     {
         Schema::table('rooms', function (Blueprint $table) {
-            $table->string('image')->nullable();
+            if (!Schema::hasColumn('rooms', 'image')) {
+                $table->string('image')->nullable();
+            }
         });
     }
 
     /**
      * Reverse the migrations.
+     *
+     * @return void
      */
-    public function down(): void
+    public function down()
     {
         Schema::table('rooms', function (Blueprint $table) {
-            $table->dropColumn('image');
+            if (Schema::hasColumn('rooms', 'image')) {
+                $table->dropColumn('image');
+            }
         });
     }
-};
+}
